refactor(network): rename get_blocks to streamed_data_protocol (#1154)
