+++ crates/papyrus_network/src/lib.rs
@@ -2,8 +2,8 @@
 /// to the [`Starknet p2p specs`]
 ///
 /// [`Starknet p2p specs`]: https://github.com/starknet-io/starknet-p2p-specs/
-pub mod get_blocks;
 pub mod messages;
+pub mod streamed_data_protocol;
 
 use starknet_api::block::{BlockHash, BlockNumber};
