+++ crates/papyrus_network/src/streamed_data_protocol/protocol_test.rs
